Separating seed generation and token processing
